fixed invalid function
terminate-user is not a command, i genuinely questioning where you got the terminate-user command from. a fix i implemented was using the exit command, idk how to compile this so you have to test it out. there is a high chance that it will break as no user is specified but its worth a shot